travis: ocamlscript + opam + topfind workaround
The ocaml toplevel needs to have the 'topfind' directory on its
include path. If using the system compiler opam hasn't got the
privileges to install in the normal place (see [ocaml/opam#45])

The normal workaround is for opam to export OCAML_TOPLEVEL_PATH
into the environment and then .ocamlinit to dereference this.
This workaround breaks for ocamlscript because it doesn't evaluate
the .ocamlinit.

We attempt to work around that by copying 'topfind' into the
usual place in the system stdlib directory.

Signed-off-by: David Scott <dave.scott@citrix.com>