Create a thin-pool and create volumes from it
In SR.create we create a thin-pool using 80% of the remaining storage
with a well-known name. We leave space free for the admin to create
additional non-thin volumes. The admin is also free to resize the thin
pool as they see fit.

In Volume.create we create our volumes from the well-known thin-pool.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>